Add config_declaration CKAN >2.10 and refactor config.py to config/
CKAN itself declares all the config options that are used throught the code base (You can see the core config declarations in the ckan/config/config_declaration.yaml file). This allows to validate the current configuration against the declaration, or check which config options in the CKAN config file are not declared (and might have no effect).
https://github.com/mjanez/ckanext-schemingdcat/issues/102